fix delete vehicle if have active parking

+++ app/Http/Controllers/Api/V1/VehicleController.php
@@ -39,6 +39,12 @@ class VehicleController extends Controller
 
     public function destroy(Vehicle $vehicle)
     {
+        if ($vehicle->hasActiveParkings()) {
+            return response()->json([
+                'errors' => ['general' => ['Can\'t remove vehicle with active parkings. Stop active parking.']],
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
+
         $vehicle->delete();
 
         return response()->noContent();
